requirements.txt
# Data Processing
pandas==2.0.3
numpy==1.24.3

# Machine Learning
torch==2.1.0 --index-url https://download.pytorch.org/whl/cpu
transformers==4.35.0
scikit-learn==1.3.0

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
mangum==0.17.0

# AWS SDK (for production)
boto3==1.29.7

# Utilities
python-dotenv==1.0.0

# Testing & Quality
pytest==7.4.0
pytest-cov==4.1.0
httpx==0.24.1
ruff==0.1.5
mypy==1.7.0
python-json-logger==2.0.7